requirements.txt
openai
duckduckgo-search
python-dotenv